requirements.txt
# Core RAG Libraries
llama-index==0.10.12
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-ollama==0.1.3

# Vector Database
faiss-cpu==1.7.4
# Open requirements.txt and add:
faiss-gpu==1.7.4    
# Note: We'll upgrade to faiss-gpu after testing

# Embeddings & NLP
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# Data Processing
pandas==2.2.0
numpy==1.26.3
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Ollama Integration
ollama==0.1.6

# Retrieval & Search
rank-bm25==0.2.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# UI Framework
streamlit==1.31.0

# Evaluation
ragas==0.1.4

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
loguru==0.7.2

# Export & Reporting
reportlab==4.0.9
markdown==3.5.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0